requirements.txt
Django==4.2.4
psycopg2-binary==2.9.10
celery==5.3.4
redis==4.6.0
gunicorn==21.1.0
requests[socks]==2.32.3
urllib3==1.26.15
django-celery-beat==2.5.0
django-modeltranslation==0.18.6
django-redis==5.4.0
django-translated-fields==0.13.0
django-jsonfield==1.4.1
sorl-thumbnail==12.10.0
django-admin-list-filter-dropdown==1.0.3
uvicorn[standard]==0.30.1
python-dateutil==2.9.0.post0
translate==3.6.1
django-unfold



pydantic==2.9.2
pydantic[email]
fastapi==0.115.4
asyncpg==0.29.0
typer==0.12.5
click==8.1.7
# uvicorn==0.15.0
aiohttp==3.10.10
backoff==1.11.1
ujson==5.10.0
json-logging==1.3.0
PyJWT==2.6.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aioredis==2.0.1
pytz==2024.2
udp-mongo-logger==1.0.2